Add ASRFader and PinSensor classes.
The ASRFader class handles the PIR sensor’s logic, including fade-in,
fade-out, and sustain.

The PinSensor abstracts the actual digitalRead() call away from this.
The primary use-case for this is to let developers swap PinSensor for a
PullupPinSensor, if you’re using a trigger switch in place of a PIR
sensor in a test rig.

The ASR Fader is more complex than the previous code, as it incorporates
a simple state machine. However, this makes it more robust. In particular
https://github.com/camp8bit/forest-lighting/issues/16 is fixed, and
the onFadeoutEnd event trigger is more reliable.