Add PatternState object to store shared state.
Previously, shared state was stored in globals. This makes it harder
to run two families of patterns concurrently and cross-fade them. So,
I’ve moved the shared state into an object.

This also means that the CRGB array is part of this state rather than
being provided in the loop() call.