Improve memory efficiency by allocating state arrays on demand
With 3 PatternState objects in play for the cross-fader, we now
allocate a lot of memory that we might not use.

To deal with this, I’ve updated the palette and activation arrays to be
allocate-on-demand, which means that the cross-fader won’t allocate
the palette and byte array unnecessarily.

It kind of complicates the pattern code, so happy to hear other
suggestions.